fix(observability,bootstrap,docs): apply round-2 review findings (D3a-1)
Round-2 review fix batch — 11 IN_SCOPE Cx1 items from D3a-1's own new code,
plus 3 OUT_OF_SCOPE items registered to backlog.

IN_SCOPE fixes (D3a-1 own code, no pre-existing pattern excuse):

ADR + archtest scope:
- F4: ADR §C — explicit Int64ObservableGauge boundary clarification
  (Observable instrument vs synchronous Gauge primitive; not in banned set
  because current GaugeVec uses Float64UpDownCounter)

Compile-time guarantees:
- F5: phases_events.go — var _ eventrouter.EventCollector compile-time
  check (parallels outbox.go's ConsumerObserver assertion; bootstrap is
  the natural location since both packages need to import each other
  for the assertion to compile without cycles)

Module organization:
- F6: autoWireOutboxConsumerCollector moved from phases_http.go to
  phases_events.go (outbox/consumer/relay are event-tier; call site
  moved from phase5BuildRouters to phase6StartEventRouter)

Documentation/godoc:
- F9: countPendingQuery comment clarifies status='pending' is bound via
  $1 parameter (no string interpolation; reviewer misread of OldestEligibleAt)
- F10: ObserveReject godoc — topic MUST be static contract ID; runtime-
  derived values risk unbounded cardinality
- F22: phaseEntries comment rewritten — false "no Gauge primitive" claim
  removed; replaced with semantic rationale (Counter for event-stream
  phase transitions vs Gauge for snapshot of phases-in-flight)
- F23: ConsumerObserver godoc — consumerGroup excluded from labels for
  cardinality reasons; pointers to OutboxConsumerCollector label set
- F27: AttachObserver godoc — single public injection path rationale
  (ConsumerBase constructed before metrics provider; constructor injection
  not viable; bootstrap.autoWireOutboxConsumerCollector wires phase 5/6)
- F28: OutboxConsumerCollector godoc — double-duty disclosure (implements
  both ConsumerObserver + ObservePendingDepth; bootstrap wires both paths
  from same instance)
- F29: cardinality quantification — {cell × topic × 2 reasons} bound
  with >100 topic monitoring threshold guidance

Test diagnostics:
- F15: TestAttrCache_ConcurrentLookupRaceSafe — LessOrEqual → LessOrEqualf
  to actually format the size/cap diagnostic message

Errcode consistency:
- F30: NewOutboxConsumerCollector + NewEventRouterCollector — caller-error
  precondition checks (nil provider, empty cellID) errcode.KindInternal
  → errcode.KindInvalid (semantically validation failures, not internal
  state errors; provider-internal Wrap calls remain KindInternal as those
  are genuine adapter failures)

F26 NOT A BUG (verified): WithRelay takes *runtimeoutbox.Relay concrete
pointer not interface; `r == nil` IS the correct check. Typed-nil only
applies to interface-typed params (the *Relay concrete cast `(*Relay)(nil)`
equals bare `nil` for concrete pointer comparison). No code change needed.

OUT_OF_SCOPE — registered to docs/backlog/cap-13-observability.md §13.5:
- BOOTSTRAP-SHUTDOWN-OUTCOME-LABEL-ALIGNMENT-01 (F17): bootstrap_shutdown_total
  outcome="success" emitter vs "clean" docs — pre-existing tech debt
  surfaced by W4 shutdown MOVE review; not introduced by D3a-1
- CONSUMER-BASE-SLOG-DEFAULT-INJECTION-01 (F16): slog.SetDefault global
  mutation pattern in kernel/outbox tests vs t.Parallel — pre-existing
  pattern; D3a-1 added one consumer but did not amplify
- ROADMAP-D3A2-UNBLOCK-ANNOTATION-01 (F33): D3a-2 roadmap row "可并行"
  wording — cosmetic, self-resolving in D3a-2 PR

Verification: golangci-lint 0 issues; go build/test clean.

Refs: D3a-1 (PR-V11-D3-R-METRIC-PACK-CORE) review round-2 fixes
