+++ docs/backlog/cap-13-observability.md
@@ -67,4 +67,7 @@
 | USER-REPO-READYZ-PROBE-01 | **UserRepository.RepoReady probe 探测 users 表 schema** — 现状: accesscore 已有 `session_store_ready` probe（探测 `sessions` 表），不覆盖 `users` 表 schema 漂移；migration 032 新增 `failed_login_count` / `last_failed_at` / `locked_until` 三列，`UpdateLockoutFields` 若列缺失会在运行期 fail，但 readyz 不感知。修复方向: 新增 `UserRepository.RepoReady` 实现，执行代表性查询（如 `SELECT failed_login_count FROM users LIMIT 1`）验证三列存在；通过 `cell.RegisterRepoReadiness(reg, "user_repo_ready", prober)` 注册（遵循 CELL-REPO-READYZ-PROBE-01 funnel）；`celltest.RunRepoReadinessConformance` 覆盖 healthy / 列删除 / mem 三路。**触发条件**: migration 漂移导致 UpdateLockoutFields 运行期失败 | feat | P3/Cx2 | 🟢 | migration 漂移导致 UpdateLockoutFields 运行期失败 | `cells/accesscore/internal/adapters/postgres/user_repo.go` + `cells/accesscore/internal/mem/user_repo.go` + `cmd/corebundle/access_module.go` | PR-618 review F7 |
 | SAFEID-UPSTREAM-FUNNEL-HARD-01 | **SafeID wire-decode 上游 Hard 升级** — 现状: PR #582 引入 `pkg/idutil.SafeID` typed funnel；下游 Hard（`SAFEID-WIREMESSAGE-USAGE-01` archtest 锁字段类型）成立，但上游"`UnmarshalEnvelope` 是 wire envelope 唯一 decode 入口"目前仅 reasoning-based（无 archtest 锁定）；修复方向: archtest caller-allowlist — 用 typeseval 检查所有"将 bytes unmarshal 到 `*outbox.WireMessage`"的调用点，要求 callee 是 `outbox.UnmarshalEnvelope`。AI-rebust 升级路径 Medium-by-archtest → Hard。触发条件: 出现第三处直接 `json.Unmarshal(bytes, &outbox.WireMessage{})` / 跨 transport 增量调用点 / archtest 周期复盘。 | arch-hard | P2/Cx2 | 🟡 | 同 触发条件 | `tools/archtest/safeid_funnel_test.go` (新增 caller archtest) + 可能 `tools/archtest/internal/typeseval/` helper | PR #582 review F2 |
 | METRICS-GAUGEVEC-UPSTREAM-HARD-01 | **GaugeVec funnel 上游 Hard 升级** — 现状: D3a-1 added `metrics.Provider.GaugeVec` interface and archtest `METRICS-GAUGEVEC-FUNNEL-01` to ban direct `prom.NewGaugeVec` / OTel `Float64UpDownCounter` calls in business packages. The downstream form-uniqueness is Hard (callee resolved via `*types.Info`), but the upstream is Medium: `prom.NewGaugeVec` is a public exported API and the Go type system cannot prevent business packages from importing prometheus directly and calling it. The archtest scope filter (excluding `adapters/prometheus/` and `adapters/otel/`) is the only caller-allowlist mechanism today. **Upgrade path**: Wrap `prom.NewGaugeVec` (and Counter / Histogram peers, for consistency) inside `adapters/prometheus/internal/promwrap/` — a Go-internal package. The `internal` mechanism prevents external imports at compile time, making upstream Hard. Same approach for the OTel side. **Risk**: blocked on refactor cost (~6h) and potentially churn across adapter-internal callsites. D3a-1 ship made the scope-filter trade-off deliberately; this entry tracks the principled upgrade. **Refs**: `tools/archtest/observability_metrics_test.go` TestGaugeVecFunnel godoc; `.claude/rules/gocell/ai-collab.md` §Funnel 双向锁评级. | arch-hard | P2/Cx2 | 🟡 | 第二个直接 NewGaugeVec 调用点出现时 | `adapters/prometheus/internal/promwrap/` (new) + `adapters/otel/internal/otelwrap/` (new) + `tools/archtest/observability_metrics_test.go` | D3a-1 ship |
+| BOOTSTRAP-SHUTDOWN-OUTCOME-LABEL-ALIGNMENT-01 | **shutdown outcome label + metric name 三源不一致** — `runtime/bootstrap/phases_shutdown.go` emits `outcome="success"` to `bootstrap_shutdown_total`, but `docs/ops/alerting-rules.md` ~line 200 documents `outcome="clean"`, and `docs/ops/graceful-shutdown-k8s.md` ~line 71 references the metric name `bootstrap_shutdown_outcome_total` (actual name is `bootstrap_shutdown_total`). Pre-existing tech debt surfaced by D3a-1 W4 shutdown_metrics MOVE review. **Upgrade path**: unify code + alerting-rules.md + graceful-shutdown-k8s.md on a single label value (`"success"` — cheaper to update 2 docs than 1 emitter + every operator dashboard). Pick during the next shutdown-related PR. **Risk**: any dashboard/alert using `outcome="clean"` or the old metric name is silently empty today; operators would have noticed if live. Self-resolving when shutdown work next occurs. | tech-debt | P2/Cx1 | 🟡 | 下次触碰 shutdown 相关代码时 | `runtime/bootstrap/phases_shutdown.go` + `docs/ops/alerting-rules.md` + `docs/ops/graceful-shutdown-k8s.md` | D3a-1 PR #589 review F17 |
+| CONSUMER-BASE-SLOG-DEFAULT-INJECTION-01 | **ConsumerBase test global slog 注入** — `kernel/outbox/consumer_base_test.go` uses `captureDefaultSlogForConsumerBase` which calls `slog.SetDefault()` globally. Pattern marks itself "Do NOT call t.Parallel()", but other tests in the same package use `t.Parallel()`. If the parallel scheduler colocates them in the same process generation, the global slog handler gets stomped — undefined log capture behavior. D3a-1 added one more test (G-05 retry-exhausted log level) that uses the same capture helper without amplifying the risk. **Upgrade path**: inject `*slog.Logger` via `ConsumerBaseConfig` instead of capturing global, allowing per-test slog without `SetDefault` mutation. **Risk**: low — pattern works today, no observed flake. Cleanup for a future kernel/outbox refactor. | tech-debt | P3/Cx1 | 🟡 | kernel/outbox 下次重构时 | `kernel/outbox/consumer_base_test.go` | D3a-1 PR #589 review F16 |
+| ROADMAP-D3A2-UNBLOCK-ANNOTATION-01 | **D3a-2 roadmap row 依赖注释过时** — After D3a-1 ship, the roadmap's D3a-2 row (line ~72 of `docs/plans/202605101839-029-master-roadmap.md`) still says "与 D3a-1 可并行（无依赖）". Post-ship it should say "D3a-1 已 ship；D3a-2 unblocked". Cosmetic — self-resolving when D3a-2 PR includes its own roadmap update. **Risk**: none (informational). | doc | P3/Cx1 | 🟡 | D3a-2 PR 自行修复 | `docs/plans/202605101839-029-master-roadmap.md` | D3a-1 PR #589 review F33 |
